develop update company (add file)

+++ src/services/mobileUserService.js
@@ -76,16 +76,13 @@ UserService.updateUser = async (userDTO, user) => {
 
     const updatedUser = await userFromDB.$query().updateByUserId(userDTO, user.sub).returning('*');
 
-    // There is no file
-    if (userDTO.efileefileid === null) {
-        // There is a file on user, then delete it
-        if (updatedUser.efileefileid !== null) {
-            await fileService.deleteFileByIdAndCreateBy(updatedUser.efileefileid, updatedUser.euserid);
-        }
+    // User didnt give a file and file existed (delete file)
+    if (updatedUser.efileefileid && userDTO.efileefileid === null) {
+        await fileService.deleteFileByIdAndCreateBy(updatedUser.efileefileid, updatedUser.euserid);
     }
-    // There is a file
-    else if (userDTO.efileefileid !== null) {
-
+    // User give a file
+    else if (userDTO.efileefileid) {
+        
         // File does not exist / not uploaded yet
         const file = await fileService.getFileByIdAndCreateBy(userDTO.efileefileid, user.sub);
